+++ docs/CHANGELOG.md
@@ -15,6 +15,11 @@ Changes are grouped as follows:
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## TBD
+### Fixed
+- The `neat.create` + `neat.to.excel(..., include_reference=True)` now correctly includes the reference data model 
+  in the Excel file.
+
 ## [0.109.2] - 31-01-**2025**
 ### Added
 - [ALPHA] Support for external modification of data model from NeatSession and its re-import